rename decoupled -> proxy (refactored Proxy implementation)
